feat: fix the issue for float64 dtype
----------
 Warning (function answer_one), your DataFrame column, named 'Energy
Supply' is of type int64, but ours is of type float64. Attempting to
convert for grading.
----------
 Warning (function answer_one), your DataFrame column, named 'Energy
Supply per Capita' is of type int64, but ours is of type float64.
Attempting to convert for grading.
----------
 Function answer_one results were not the same as the solution results.
Iterating through the dataframes we found that there were differences in
the following columns: dict_keys(['Citations', 'Citations per document',
'2008', '2013', '2007', 'Energy Supply', 'Energy Supply per Capita',
'Rank', '2011', 'Self-citations', '2009', '2010', '2015', '2012',
'Citable documents', '% Renewable', 'Documents', '2006', 'H index',
'2014']).
----------
 Function answer_one was answered incorrectly, 0.2 points were not
awarded. The shape of the DataFrame was correct, and the column names
were correct, but equality testing failed.