docs: add tutorial for NN cost function calculation
The y matrix should not be used directly.  With multiple classes the
y vector having values from 1 to 10.   However, the calculation of cost
function with y values are either 0 or 1.  Therefore, the vector should
convert into auxiliary matrix with 10 column that each column containing
values 0 & 1 and mapping the columns to different classes.